Move glam and macaw to their own files
